packages/gossiplog: rename src/gossiplog.ts to src/GossipLog.ts
