+++ packages/gossiplog/package.json
@@ -41,6 +41,7 @@
     "@ipld/dag-json": "^10.1.3",
     "@ipld/schema": "^4.1.10",
     "@libp2p/interface-connection": "^5.1.1",
+    "@libp2p/interface-libp2p": "^3.2.0",
     "@libp2p/interface-peer-id": "^2.0.2",
     "@libp2p/interface-pubsub": "^4.0.1",
     "@libp2p/interface-registrar": "^2.0.12",
