+++ packages/gossiplog/src/GossipLog.ts
